+++ src/components/menu/Menu.vue
@@ -1,0 +1,20 @@
+<template>
+    <div :class="{ menu: true, 'is-dark': isDark }">
+        <slot name="title" />
+        <div class="divider" />
+        <ul class="menu-list">
+            <slot />
+        </ul>
+    </div>
+</template>
+
+<script>
+
+    export default {
+        name: 'BMenu',
+        props: {
+            isDark: Boolean
+        }
+    }
+
+</script>